requirements.txt
Flask==2.3.2
flask_cors==3.0.10
waitress==2.1.2
pydantic==2.3.0
pydantic_core==2.6.3
pyjwt==2.8.0
scikit-learn==1.3.0
sentence_transformers==2.2.2
openai==0.28.0
pandas==2.0.3
rapidocr_onnxruntime==1.3.6
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
filetype==1.2.0
PyMuPDF==1.22.5